gsw: move format_age_detailed alongside its short-form sibling
format_age lives in age.rs; the two-unit variant was tucked into
render.rs purely as an accident of where the first caller sits. Group
them together (and their tests).
